backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
supabase
google-generativeai
pydantic
pydantic-settings
python-jose[cryptography]
python-multipart
httpx
agno
aiohttp